refactor(phase-3): extract OrderBoard + useBatchPromote (de-dup grids)
FoodOrderGrid and DrinkOrderGrid were ~95% duplicated (~360 lines
each — every batch-promote / long-press / dialog-gate fix had to be
made twice). Both collapse to ~25-line wrappers around a parameterized
OrderBoard, with the selection state machine + optimistic fan-out
lifted into useBatchPromote. Closes the BLOCK from the Toss FE review.
